requirements.txt
asgiref==3.4.1
bencodepy==0.9.5
click==8.0.1
dotmap==1.3.24
fastapi==0.68.1
h11==0.12.0
mongo==0.2.0
pydantic==1.8.2
pymongo==3.12.0
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
